(chore) - save size and move tests
