+++ test/browser/hooks/useImperativeHandle.test.js
@@ -1,6 +1,5 @@
-import { createElement as h, render } from 'preact';
-import { setupScratch, teardown } from '../../../test/_util/helpers';
-import { useImperativeHandle, useRef } from '../../src';
+import { createElement as h, render, useImperativeHandle, useRef } from '../../../src';
+import { setupScratch, teardown } from '../../_util/helpers';
 
 /** @jsx h */
