+++ test/browser/hooks/useLayoutEffect.test.js
@@ -1,8 +1,7 @@
 import { act } from 'preact/test-utils';
-import { createElement as h, render } from 'preact';
-import { setupScratch, teardown } from '../../../test/_util/helpers';
+import { createElement as h, render, useLayoutEffect } from '../../../src';
+import { setupScratch, teardown } from '../../_util/helpers';
 import { useEffectAssertions } from './useEffectAssertions.test';
-import { useLayoutEffect } from '../../src';
 
 /** @jsx h */
