+++ test/browser/hooks/useReducer.test.js
@@ -1,7 +1,6 @@
 import { setupRerender } from 'preact/test-utils';
-import { createElement as h, render } from 'preact';
-import { setupScratch, teardown } from '../../../test/_util/helpers';
-import { useReducer } from '../../src';
+import { createElement as h, render, useReducer } from '../../../src';
+import { setupScratch, teardown } from '../../_util/helpers';
 
 /** @jsx h */
