+++ test/browser/hooks/useRef.test.js
@@ -1,6 +1,5 @@
-import { createElement as h, render } from 'preact';
-import { setupScratch, teardown } from '../../../test/_util/helpers';
-import { useRef } from '../../src';
+import { createElement as h, render, useRef } from '../../../src';
+import { setupScratch, teardown } from '../../_util/helpers';
 
 /** @jsx h */
